fix: Update clients for PageHeader proto changes and saga API
Runtime images:
- Add COPY crates/ to Containerfile for trivial-delegation macro

C++ client:
- Use page->mutable_header()->set_sequence() instead of page->set_sequence()

C# client:
- Cast uint to int for Handle method sequence parameter

Go client:
- Update CommandBuilder to use PageHeader with SequenceType oneof
- Migrate saga handlers from two-phase (Prepare/Execute) to single-phase (Handle)
- Remove SagaBase.Prepares() method (sagas are now stateless translators)
- Update OOSaga interface to use Handle instead of PrepareDestinations/Execute

Go examples:
- Update sagas to use new Handle-only API
- Use angzarr_deferred for sequence stamping (framework handles it)

Python:
- Formatting changes from black/ruff

+++ client/csharp/Angzarr.Client/Router/CommandHandlerRouter.cs
@@ -114,7 +114,7 @@ public class CommandHandlerRouter<TState, THandler>
         }
 
         // Execute handler
-        var resultBook = _handler.Handle(commandBook, commandAny, state, seq);
+        var resultBook = _handler.Handle(commandBook, commandAny, state, (int)seq);
 
         return new Angzarr.BusinessResponse { Events = resultBook };
     }
